+++ react_typescript/src/type/TabType.ts
@@ -1,0 +1,1 @@
+export type TabTypes = "all" | "onlyStudents" | "onlyMentors";
